ロジックをhooksに移動
